Arreglos de warnings

+++ Kernel/include/time.h
@@ -1,8 +1,9 @@
 #ifndef _TIME_H_
 #define _TIME_H_
+#include <stdint.h>
 
 void timerHandler();
-int ticksElapsed();
+uint64_t ticksElapsed();
 int secondsElapsed();
 
 #endif
